+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Änderungen
 ----------
 v0.14.15 (??.??.2015)
+ * possiblity to configure allowed and disallowed file extensions for FAL TCA fields in $field['config']['allowedFileExtensions'] and $field['config']['disallowedFileExtensions']
  
 v0.14.14 (23.03.2015)
  * new wrapper class for network related TYPO3 functions like cmpIP